requirements.txt
albumentations==1.4.3  # 1.3.0升级albumentations到最新
einops==0.3.0
kornia==0.4.1
loguru==0.5.3
matplotlib==3.8.4    # 初始为3.3.4
numpy==1.26.4        # NumPy 2.1.3降级NumPy以兼容其他库
opencv_python==4.9.0.80
Pillow==11.0.0
pyproj==3.6.1
PyYAML==6.0.2
scikit_learn==1.5.2  # 1.4.0升级scikit-learn到支持新NumPy的版本
scipy==1.14.1
setuptools==68.2.2
scikit-image
timm==1.0.11
torch==2.2.1 #原为torch==2.2.1+cu121下行同理
torchvision==0.17.1
tqdm==4.66.1
transformers==4.37.0
#xformers # Optional, used for memefficient attention
